sfuaniweb/mysite/requirements.txt
asgiref==3.2.7
Django==3.0.5
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-summernote==0.8.11.6
pytz==2019.3
sqlparse==0.3.1
whitenoise==5.1.0
